+++ projects/net.splitcells.dem/src/main/java/net/splitcells/dem/resource/communication/interaction/LogMessageI.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR MIT
  */
-package net.splitcells.dem.resource.host.interaction;
+package net.splitcells.dem.resource.communication.interaction;
 
 import net.splitcells.dem.data.set.list.List;
 import net.splitcells.dem.object.Discoverable;