+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/optimization/meta/hill/climber/FunctionalHillClimber.java
@@ -10,7 +10,7 @@
  */
 package net.splitcells.gel.solution.optimization.meta.hill.climber;
 
-import net.splitcells.dem.resource.host.interaction.LogLevel;
+import net.splitcells.dem.resource.communication.interaction.LogLevel;
 import net.splitcells.dem.data.set.list.List;
 import net.splitcells.gel.solution.SolutionView;
 import net.splitcells.gel.rating.framework.Rating;