+++ projects/net.splitcells.website.server/src/main/java/net/splitcells/website/server/renderer/ProjectRendererI.java
@@ -15,7 +15,7 @@ import net.splitcells.dem.lang.Xml;
 import net.splitcells.dem.lang.namespace.NameSpaces;
 import net.splitcells.dem.lang.perspective.Perspective;
 import net.splitcells.dem.resource.Files;
-import net.splitcells.dem.resource.host.interaction.LogLevel;
+import net.splitcells.dem.resource.communication.interaction.LogLevel;
 import net.splitcells.website.Validator;
 import net.splitcells.website.server.renderer.extension.ExtensionMerger;
 import net.splitcells.website.server.renderer.extension.commonmark.CommonMarkRenderer;